+++ pgsrip/media.py
@@ -157,7 +157,7 @@ class Pgs:
 
     @property
     def srt_path(self) -> MediaPath:
-        return self.media_path.translate(number=0, extension='srt')
+        return self.media_path.translate(extension='srt')
 
     @property
     def items(self) -> list[PgsSubtitleItem]: